Move routing redirect spec so it works
